+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -29,6 +29,7 @@
 - Parameter `useTrainingLabels` is now required for `FormTrainingClient.StartTraining`.
 - Protected constructors have been removed from `Operation` types, such as `TrainingOperation` or `RecognizeContentOperation`.
 - `USReceipt`, `USReceiptItem`, `USReceiptType` and `FormField{T}` types removed. Information about a `RecognizedReceipt` must now be extracted from its `RecognizedForm`.
+- `ReceiptLocale` removed from `RecognizedReceipt`.
 
 ### New Features